Reorganized globals.
