+++ impls/cpp.2/Makefile
@@ -13,36 +13,40 @@ $(BIN)/lineedit.o $(BIN)/reader.o $(BIN)/types.o $(BIN)/printer.o \
 $(BIN)/parse_numbers.o $(BIN)/parse_sequences.o $(BIN)/parse_reader_macros.o \
 -o $(BIN)/step1_read_print
 
-step2_eval: step2_eval.cpp lineedit reader printer types parse_numbers parse_sequences parse_reader_macros eval env apply
+step2_eval: step2_eval.cpp lineedit reader printer types parse_numbers parse_sequences parse_reader_macros eval env apply globals
 	$(CC) $(FLAGS) -I$(INC) $(LIBS) step2_eval.cpp \
 $(BIN)/lineedit.o $(BIN)/reader.o $(BIN)/types.o $(BIN)/printer.o \
 $(BIN)/parse_numbers.o $(BIN)/parse_sequences.o $(BIN)/parse_reader_macros.o \
-$(BIN)/eval.o $(BIN)/env.o $(BIN)/apply.o \
+$(BIN)/eval.o $(BIN)/env.o $(BIN)/apply.o $(BIN)/globals.o \
 -o $(BIN)/step2_eval
 
 
-step3_env: step3_env.cpp lineedit reader printer types parse_numbers parse_sequences parse_reader_macros eval env apply
+step3_env: step3_env.cpp lineedit reader printer types parse_numbers parse_sequences parse_reader_macros eval env apply globals
 	$(CC) $(FLAGS) -I$(INC) $(LIBS) step3_env.cpp \
 $(BIN)/lineedit.o $(BIN)/reader.o $(BIN)/types.o $(BIN)/printer.o \
 $(BIN)/parse_numbers.o $(BIN)/parse_sequences.o $(BIN)/parse_reader_macros.o \
-$(BIN)/eval.o $(BIN)/env.o $(BIN)/apply.o \
+$(BIN)/eval.o $(BIN)/env.o $(BIN)/apply.o $(BIN)/globals.o \
 -o $(BIN)/step3_env
 
 
 
-step4_if_fn_do: step4_if_fn_do.cpp lineedit reader printer types parse_numbers parse_sequences parse_reader_macros eval env apply
+step4_if_fn_do: step4_if_fn_do.cpp lineedit reader printer types parse_numbers parse_sequences parse_reader_macros eval env apply globals
 	$(CC) $(FLAGS) -I$(INC) $(LIBS) step4_if_fn_do.cpp \
 $(BIN)/lineedit.o $(BIN)/reader.o $(BIN)/types.o $(BIN)/printer.o \
 $(BIN)/parse_numbers.o $(BIN)/parse_sequences.o $(BIN)/parse_reader_macros.o \
-$(BIN)/eval.o $(BIN)/env.o $(BIN)/apply.o \
+$(BIN)/eval.o $(BIN)/env.o $(BIN)/apply.o $(BIN)/globals.o \
 -o $(BIN)/step4_if_fn_do
 
+
+globals: globals.cpp types env apply
+	$(CC) $(FLAGS) -I$(INC) $(LIBS) -c globals.cpp -o $(BIN)/globals.o
 
 eval: eval.cpp types env apply
 	$(CC) $(FLAGS) -I$(INC) $(LIBS) -c eval.cpp -o $(BIN)/eval.o
 
 apply: apply.cpp types env
 	$(CC) $(FLAGS) -I$(INC) $(LIBS) -c apply.cpp -o $(BIN)/apply.o
+
 env: env.cpp types
 	$(CC) $(FLAGS) -I$(INC) $(LIBS) -c env.cpp -o $(BIN)/env.o
