+++ impls/cpp.2/includes/env.h
@@ -66,24 +66,24 @@ protected:
 class Env_Primitive: public Env_Symbol
 {
 public:
-    Env_Primitive(MalPtr s, std::function<TokenVector(TokenVector)>& f, int a): Env_Symbol(s), fn(f) {n_ary = a;};
+    Env_Primitive(MalPtr s, Procedure p, int a): Env_Symbol(s), procedure(p) {n_ary = a;};
     virtual Env_Element_Type type() {return ENV_PRIMITIVE;};
     virtual TokenVector apply(TokenVector& args);
 
 protected:
-    std::function<TokenVector(TokenVector)> fn;
+    Procedure procedure;
 };
 
 
 class Env_Procedure: public Env_Symbol
 {
 public:
-    Env_Procedure(MalPtr s, TokenVector& f, int a): Env_Symbol(s), fn(f) {n_ary = a;};
+    Env_Procedure(MalPtr s, MalPtr p, int a): Env_Symbol(s), procedure(p) {n_ary = a;};
     virtual Env_Element_Type type() {return ENV_PROCEDURE;};
     virtual TokenVector apply(TokenVector& args);
-
+    virtual MalPtr fn() {return procedure;};
 protected:
-    TokenVector fn;
+    MalPtr procedure;
 };
